renamed and extracted some interfaces
