+++ codemod/output/S3PartInterface.ts
